==> 04_post_vsp_dataprocessing.ipynb <==
import pandas as pd
import numpy as np
import pickle
import os
import matplotlib.pyplot as plt
import pyspark.sql.functions as F
import pyspark.sql.types as T
import pickle
import re
import pandas as pd
from scipy import sparse, io
import matplotlib.pylab as plt
import numpy as np
from glob import glob
from pyspark.sql import SparkSession
from tqdm import tqdm
from collections import defaultdict, Counter
import sys

os.environ['PYSPARK_PYTHON'] = sys.executable
os.environ['PYSPARK_DRIVER_PYTHON'] = sys.executable

RAW_TWEETS_PATH = "/data/dnc2020/raw_tweets/"
PARQUET_TWEET_PATH = '/data/navid/processed_tweets'
PARQUET_USER_PATH = "/data/navid/processed_users"
PARQUET_RTNET_PATH = "/data/kenny/processed_rt_network"

## We use spark to analyze the data efficiently. You may opt for a different format
## Also note we here assume a machine with a reasonably large number of cores (16) 
## and RAM (at least 32 GB). Adjust below accordingly
spark = (
    SparkSession
    .builder
    .master("local[{}]".format(16))
    .config("spark.driver.memory", "{}g".format(50))
    .config("spark.driver.maxResultSize", f"{10}g")
    .getOrCreate()
)


# Write out subset of tweets from Bernie Users
 

all_tweets_df = spark.read.parquet(os.path.join(PARQUET_TWEET_PATH, '*', '*'))

clustering_res = spark.read.csv("/data/dnc2020/bernie_vsp.csv",sep=" ", header=True)

clustering_res.show(5,False)

cluster_tweets = clustering_res.join(all_tweets_df, ['uid'])

cluster_tweets.write.parquet("/data/dnc2020/tweets_clustering",mode="overwrite")

# Evaluation: RTs of Candidates (generate data)

accts = [("939091", "JoeBiden"),
("1193981789787824129", "Mike2020"),
("226222147", "PeteButtigieg"),
("426028646", "JohnDelaney"),
("26637348", "TulsiGabbard"),
("33537967", "amyklobuchar"),
("61636488", "DevalPatrick"),
("216776631", "BernieSanders"),
("949934436", "TomSteyer"),
("357606935", "ewarren"),
("2228878592", "AndrewYang"),
("4797361833", "TomPerez"),
("1108472017144201216", "TrumpWarRoom"),
("818910970567344128", "VP45")]

spark = (
    SparkSession
    .builder
    .master("local[{}]".format(16))
    .config("spark.driver.memory", "{}g".format(50))
    .config("spark.driver.maxResultSize", f"{20}g")
    .getOrCreate()
)

trimmed_rt_net = spark.read.parquet("data/trimmed_rtnet/")

acct_df = spark.createDataFrame(accts, schema=T.StructType([
                    T.StructField('rt_uid', T.StringType(), True), 
                   T.StructField('sn', T.StringType(), True)]))

trimmed_rt_net = trimmed_rt_net.join(acct_df, "rt_uid")

trimmed_rt_net.write.parquet("rtnet_eval",mode='overwrite')

# Evaluation: Most Common Identities Per Cluster

user_desc_df = spark.read.parquet("/data/dnc2020/user_descript")

from personal_identifiers import *

@F.udf(returnType=T.ArrayType(T.StringType()))
def extract_pid(bio):
    if bio is None:
        return ""
    return find_identifiers_simple(bio)[0]

user_to_pi = (user_desc_df
 .withColumn('pi', extract_pid('description'))
 .select(F.col("id"),F.explode(F.col("pi")))
 .dropDuplicates(subset=['id','col'])
)

user_to_pi.write.parquet("/data/dnc2020/user_pi_net", mode="overwrite")

user_to_pi = spark.read.parquet("/data/dnc2020/user_pi_net/")

clustering_res = spark.read.csv("bernie_vsp_yz.csv",sep=" ",
                                header=True)
clustering_res.show(5,False)

user_to_pi.show(5, False)

mg = user_to_pi.join(clustering_res,
                     user_to_pi.id == clustering_res.uid)

mg.groupby(["col","grp","utype"]).count().write.parquet("bern_pi",mode="overwrite")

spark.stop()

df = pd.read_parquet("data/bern_pi/")

df.head()

import pandas as pd
import numpy as np

"""Fx to get weighted log odds"""
def get_weighted_log_odds(dataFrame, groupSet, featureName, nCount):
    """
    Fx to get weighted log odds ratio of features depending on their grouping.
    Adapted from https://github.com/juliasilge/tidylo/blob/main/R/bind_log_odds.R
    Input is a dataframe that has columns of your set (groups you are comparing to one another),
    feature (the thing you want to see the weighted log odds of) and the count (n) of
    features per set. This is flexible enough to get weighted log odds of any set, feature
    and count combination.
    """
    alphas = dataFrame.groupby(featureName).sum(nCount).reset_index()[[featureName, nCount]].rename(columns={nCount:'alpha'})  # Get total # of features across whole dataset
    dataFrame_II = dataFrame.merge(alphas, on= featureName, how = 'left')
    dataFrame_II['y_wi'] = dataFrame_II[nCount] + dataFrame_II['alpha']  # "y_wi is the pseudo count of word w in group i"
    y_ws = dataFrame_II.groupby(featureName).sum('y_wi').reset_index()[[featureName, 'y_wi']].rename(columns={'y_wi': 'y_w'})  # "y_w is the total count of word w"
    # Actually is sort of an exaggerated count ¯\_(ツ)_/¯
    dataFrame_III = dataFrame_II.merge(y_ws, on=featureName, how='left')
    # Do exaggerated count per set:
    n_is = dataFrame_II.groupby(groupSet).sum('y_wi').reset_index()[[groupSet, 'y_wi']].rename(columns={'y_wi': 'n_i'})
    pseudo_counts = dataFrame_III.merge(n_is, on= groupSet, how = 'left')  # Combine everything together! Now ready to do weighted log odds calc
    pseudo_counts['omega_wi'] = pseudo_counts['y_wi'] / (pseudo_counts['n_i'] - pseudo_counts['y_wi'])  # # odds in group i
    pseudo_counts['omega_w'] = pseudo_counts['y_w'] / (sum(pseudo_counts['y_wi']) - pseudo_counts['y_w'])  # overall odds
    pseudo_counts['delta_wi'] = (np.log(pseudo_counts['omega_wi'])) - (np.log(pseudo_counts['omega_w']))   # eqn 15 << not sure what this means, it's in OG R code
    pseudo_counts['sigma2_wi'] = 1 / pseudo_counts['y_wi'] + 1 / pseudo_counts['y_w']           # eqn 18
    pseudo_counts['zeta_wi'] = pseudo_counts['delta_wi'] / (np.sqrt(pseudo_counts['sigma2_wi']))       # eqn 21
    return pseudo_counts.rename(columns={'zeta_wi': 'log_odds_weighted', 'delta_wi': 'log_odds'}).sort_values('log_odds_weighted', ascending = False)

df.columns = ['col','grp','utype','n']

df_res = get_weighted_log_odds(df[df.utype == "influencer"], "grp","col","n")

df_res.grp.unique()

for g in df_res.grp.unique():
    print(g + " \t " + ", ".join(df_res[(df_res.grp == g)  ].sort_values(
        "log_odds_weighted",ascending=False).iloc[:25].col) + "")
